# Fix off-heap byte vector scoring at query time