@msokolov 2025-07-03
I would be curious to see if the results hold up for 768d vectors.  Maybe 300 not being divisible by 8 is problematic, although I don't believe it. But ... the results look so weird, I just want to try some other random thing and see if it holds up?